Add a piecemeal upgrade scenario
This test does much the same thing that the regular grizzly2havana scenario
does, except this one does one service at a time. Effectively:

for service in keystone glance cinder nova:
  1. Shut down $service
  2. Upgrade packages for $service
  3. Upgrade the database schema for $service
  4. Start up $service
  5. Create a guest and make sure it's happy

If everything works, the result of the test is five running and healthy guests,
one created in the original environment, and one for each service upgrade step
along the way.